requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-allauth==65.15.0
django-cors-headers==4.9.0
djangorestframework==3.17.0
idna==3.11
mercadopago==2.3.0
pillow==12.1.1
PyJWT==2.12.1
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
